requirements.txt
anyio==4.11.0
backoff==2.2.1
certifi==2025.10.5
cffi==2.0.0
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
IMAPClient==3.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.23
Pygments==2.19.2
python-dotenv==1.2.1
python-telegram-bot==22.5
pyzmail36==1.0.5
rich==14.2.0
setuptools==80.9.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
websocket==0.2.1
websockets==15.0.1
zope.event==6.0
zope.interface==8.0.1
